requirements.txt
# Web Framework
Flask==3.0.0
Flask-Login==0.6.3

# Database
# SQLite is built into Python, no package needed

# Social Media APIs
tweepy==4.14.0  # Twitter/X
praw==7.7.1     # Reddit

# Email
sendgrid==6.11.0

# Authentication & Security
bcrypt==4.1.1
python-dotenv==1.0.0

# Data Processing
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Code Quality
black==23.12.1
flake8==6.1.0

# WSGI Server (for production)
gunicorn==21.2.0
